+++ src/parac/preprocessor/ctx.py
@@ -10,6 +10,9 @@ from os import PathLike
 from typing import Dict, Union, List, TYPE_CHECKING, Tuple
 import antlr4
 
+
+from parac import (FailedToProcessError, ParserError, LexerError,
+                   ParaCSyntaxErrorCollection, ParacCompilerError)
 from ..abc import ProgramRunContext, FileRunContext
 from .logic_stream import PreProcessorStream
 from .__main__ import PreProcessor, PreProcessorProcessResult
@@ -149,19 +152,19 @@ class ProgramPreProcessorContext(ProgramRunContext):
         ...
 
     async def process_program(
-            self, enable_out: bool
+            self, log_errors_and_warnings: bool
     ) -> PreProcessorProcessResult:
         """
         Processes this instance and generates the logic streams required
         for generating the finished code.
 
-        :param enable_out: If set to True errors, warnings and info will be
-        logged onto the console using the local logger instance. If an
-        exception is raised or error is encountered, it will be reraised with
-        the FailedToProcessError.
+        :param log_errors_and_warnings: If set to True errors, warnings and
+         info will be logged onto the console using the local logger instance.
+         If an exception is raised or error is encountered, it will be reraised
+         with the FailedToProcessError.
         :returns: A PreProcessorProcessResult instance
         """
-        await self.parse_entry_file(enable_out)
+        await self.parse_entry_file(log_errors_and_warnings)
 
         ...  # TODO! Run listener for every file
 
@@ -171,22 +174,23 @@ class ProgramPreProcessorContext(ProgramRunContext):
     async def get_stream_and_parse(
             self,
             file_path: Union[str, PathLike],
-            enable_out: bool
+            log_errors_and_warnings: bool
     ) -> FilePreProcessorContext:
         """
         Gets a FileStream, converts it to a string stream and parses it
         returning the resulting FilePreProcessorContext
 
         :param file_path: Path to the file
-        :param enable_out: If set to True errors, warnings and info will be
-        logged onto the console using the local logger instance. If an
-        exception is raised or error is encountered, it will be reraised with
-        the FailedToProcessError.
+        :param log_errors_and_warnings: If set to True errors, warnings and
+         info will be logged onto the console using the local logger instance.
+         If an exception is raised or error is encountered, it will be reraised
+         with the FailedToProcessError.
         :returns: The FilePreProcessorContext instance for the file
         """
         from ..compiler import ParacCompiler
-        from ..util import (get_file_stream, get_relative_file_name,
-                            get_input_stream)
+        from ..util import (
+            get_file_stream, get_relative_file_name, get_input_stream
+        )
 
         file_stream = get_file_stream(file_path, self.encoding)
         relative_file_name = get_relative_file_name(
@@ -200,31 +204,41 @@ class ProgramPreProcessorContext(ProgramRunContext):
             name=file_stream.name
         )
         return await self.parse_single_file(
-            stream, relative_file_name, enable_out
+            stream, relative_file_name, log_errors_and_warnings
         )
 
     @staticmethod
     async def parse_single_file(
             stream: antlr4.InputStream,
             relative_file_name: str,
-            enable_out: bool,
+            log_errors_and_warnings: bool,
     ) -> FilePreProcessorContext:
         """
         Parses a single file and generates the FilePreProcessorContext
 
         :param stream: The Antlr4 InputStream which represents a string stream
         :param relative_file_name: Relative name of the file (fetch-able
-        using get_relative_file_name)
-        :param enable_out: If set to True errors, warnings and info will be
-        logged onto the console using the local logger instance. If an
-        exception is raised or error is encountered, it will be reraised with
-        the FailedToProcessError.
+         using get_relative_file_name)
+        :param log_errors_and_warnings: If set to True errors, warnings and
+         info will be logged onto the console using the local logger instance.
+         If an exception is raised or error is encountered, it will be reraised
+         with the FailedToProcessError.
         :returns: The generated FilePreProcessorContext instance
+        :raises FailedToProcessError: If log_errors_and_warnings is True and
+         an exception is encountered. The logs of the exception will be printed
+         onto the console.
         """
         from .listener import Listener
 
-        antlr4_file_ctx = await PreProcessor.parse(stream, enable_out)
-        listener = Listener(antlr4_file_ctx, stream, relative_file_name)
-        await listener.walk_and_process_directives(enable_out)
+        try:
+            antlr4_file_ctx = await PreProcessor.parse(stream, log_errors_and_warnings)
+            listener = Listener(antlr4_file_ctx, stream, relative_file_name)
+            await listener.walk_and_process_directives(log_errors_and_warnings)
 
-        return listener.file_ctx
+            return listener.file_ctx
+        except (LexerError, ParserError, ParaCSyntaxErrorCollection,
+                ParacCompilerError) as e:
+            if log_errors_and_warnings:
+                raise FailedToProcessError(exc=e) from e
+            else:
+                raise e