+++ src/parac/util/__init__.py
@@ -7,3 +7,4 @@ from .decorators import *
 from .pathtools import *
 from .stream import *
 from .strtools import *
+from .parser_ctx_tools import *